make ParticipationModel a Common DTO Model
