+++ src/NTS.Application/DataTransferObjects/Models/CoreModel.cs
@@ -3,11 +3,10 @@ using NTS.Domain.Core.Aggregates;
 using NTS.Domain.Core.Aggregates.Participations;
 using NTS.Domain.Core.Objects;
 using NTS.Domain.Enums;
-using NTS.Domain.Objects;
 using static NTS.Application.DataTransferObjects.Models.CommonModel;
-using static NTS.Application.DataTransferObjects.Models.SetupModel;
 
 namespace NTS.Application.DataTransferObjects.Models;
+
 public class CoreModel
 {
     public class CompetitionModel
@@ -29,47 +28,6 @@ public class CoreModel
         public Competition ToDomain()
         {
             return new Competition(Name, Ruleset, Type);
-        }
-    }
-
-    public class CombinationModel
-    {
-        public static CombinationModel Create(Combination combination)
-        {
-            return new CombinationModel
-            {
-                Number = combination.Number,
-                Distance = combination.Distance,
-                MinAverageSpeed = combination.MinAverageSpeed,
-                MaxAverageSpeed = combination.MaxAverageSpeed,
-                Athlete = AthleteModel.Create(combination.Athlete),
-                Horse = HorseModel.Create(combination.Horse),
-            };
-        }
-
-        public int Number { get; init; }
-        public string Distance { get; init; } = default!;
-        public double? MinAverageSpeed { get; init; }
-        public double? MaxAverageSpeed { get; init; }
-        public AthleteModel Athlete { get; init; } = default!;
-        public HorseModel Horse { get; init; } = default!;
-
-        public Combination ToDomain()
-        {
-            var athlete = new Athlete(Athlete.ToSetupDomain());
-            var horse = new Horse(Horse.ToSetupDomain());
-            var minSpeed = Speed.Create(MinAverageSpeed);
-            var maxSpeed = Speed.Create(MaxAverageSpeed);
-            return new Combination(
-                RandomHelper.GenerateUniqueInteger(),
-                Number,
-                athlete,
-                horse,
-                athlete.Club,
-                Distance,
-                minSpeed,
-                maxSpeed
-            );
         }
     }
 
@@ -216,48 +174,6 @@ public class CoreModel
         }
     }
 
-    public class ParticipationModel
-    {
-        public static ParticipationModel Create(Participation participation)
-        {
-            var total = participation.GetTotal();
-
-            return new ParticipationModel
-            {
-                Category = participation.Category,
-                Competition = CompetitionModel.Create(participation.Competition),
-                Combination = CombinationModel.Create(participation.Combination),
-                Phases = participation.Phases.Select(PhaseModel.Create).ToArray(),
-                Total = total == null ? null : TotalModel.Create(total),
-                Eliminated =
-                    participation.Eliminated == null ? null : EliminatedModel.Create(participation.Eliminated),
-            };
-        }
-
-        public ParticipationCategory Category { get; init; } = default!;
-        public CompetitionModel Competition { get; init; } = default!;
-        public CombinationModel Combination { get; init; } = default!;
-        public PhaseModel[] Phases { get; init; } = default!;
-        public TotalModel? Total { get; init; }
-        public EliminatedModel? Eliminated { get; init; }
-
-        public Participation ToDomain()
-        {
-            var competition = Competition.ToDomain();
-            var combination = Combination.ToDomain();
-            var phases = Phases.Select(x => x.ToDomain());
-            var eliminated = Eliminated?.ToDomain();
-            return new Participation(
-                RandomHelper.GenerateUniqueInteger(),
-                Category,
-                competition,
-                combination,
-                new(phases),
-                eliminated
-            );
-        }
-    }
-
     public class RankingEntryModel
     {
         public static RankingEntryModel Create(RankingEntry rankingEntry)
@@ -276,7 +192,7 @@ public class CoreModel
 
         public RankingEntry ToDomain()
         {
-            var participation = Participation.ToDomain();
+            var participation = Participation.ToCoreDomain();
             return new RankingEntry(RandomHelper.GenerateUniqueInteger(), participation, Rank, IsNotRanked);
         }
     }
